[#14]refactor: headerView로 재사용 가능한 View 생성
